added-jam-variety service and component

+++ Data/Variety.cs
@@ -3,5 +3,5 @@ namespace BlazorApp.Data;
 public sealed record Variety
 {
     public string Name { get; init; }
-    public decimal JamMultiplier { get; init; }
+    public double JamMultiplier { get; init; }
 }
